Function to set txpower is added.
set_txpower can be used to set transmit power. Disabling radio is not possible
yet. The semantics of setting radio value to 0 in linux is different from
NDIS it seems. The equivalent NDIS call is DISASSOCIATE, but that also
disconnects from current AP; then how do we reassociate? Should we simply
reasssociate? What if the user changes other parameters meanwhile. Lot of
issues. So for now, we just let the txpower to be 0, but that doesn't seem
to change anything, which is not harmful.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@114 388e3070-e2af-4704-926e-a220f121eef5