More WEP fixes.
Change key only if one is given. Don't assume security mode to be either
open or restricted; check what is given in both cases. Assume restricted
mode if neither is specified, which seems to work for most people, but some
seem to need open mode; give a warning in those cases.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@128 388e3070-e2af-4704-926e-a220f121eef5