Preliminary support for Realtek chip based cards is added.
These cards don't work yet in Managed mode. With some drivers, they seem
to work in Ad-Hoc mode. With this, others can probably test and debug to
get them to work. For proper working, parser needs to be changed (the current
parser fails on inf files from all vendors of this chip). For now, much
of the additions are in rt8180.c file (so it doesn't conflict with the
rest of the code). Once things work, this functionality can be moved to
either ndis.c or misc_funcs.c.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@213 388e3070-e2af-4704-926e-a220f121eef5