Power management works with suspend/resume from memory (S3) with 2.6 now.
We need to restore power state with the new driver model. With this now
resume works, but the interface for some reason is not useful. But after
resuming, ifdown/up brings it up properly (earlier after resume no matter
what is done, it wouldn't configure the interface). More work is needed
to configure the card without having to do ifdown/up after resume.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@254 388e3070-e2af-4704-926e-a220f121eef5