During initialization call pci_set_power_state and restore_state so the driver works after suspend-to-ram.
If ndiswrapper was not loaded but the machine is suspended then resumed,
the card is in powered down state. So set the power state to 0 and
resume_state with NULL so it is initialized properly. This error will
probably fix the problem with card not getting initialized that people
have complained.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@269 388e3070-e2af-4704-926e-a220f121eef5