spinlocks in ndis and linux are different, so treat ndis spinlocks as pointers and allocate space for them in KeInitializeSpinLock. However, there is no corresponding 'remove spinlock', so we keep track of these allocations and deallocate them when unloading the driver. All places where we do these kind of allocation should use the new allocation routine so memory is not leaked.
Allocations are implemented using linked list. Using list_head instead would
be simpler, but since we have to map the pointers (to be freed) back to
list entry and then delete the list entry, it would need two traversals
of the list, which is inefficient. If it happens that the number of entries
allocated/freed are not much, then we can reimplement using list_heads.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@308 388e3070-e2af-4704-926e-a220f121eef5