Wrapper function tables are now part of the file in which the functions are defined. Thus, no need to repeat them in another file.
There is now simpler interface (using macros) to enter the wrapper functions,
so there is less chance of errors and easy to maintain (the table can be easily
generated by ctags+awk).

Moved some functions from ntoskernel.c to misc_funcs.c to keep ntoskernel.c
simple so it is easy to maintain.

Added wrap_kmalloc/wrap_kfree functions so that memory allocated for
structures can be freed when module is being removed. This is required
for KeInitializeSpinLock lock function which doesn't have corresponding
destroy function.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@396 388e3070-e2af-4704-926e-a220f121eef5