Some drivers don't use _EX version of BSSID structure etc. Accessing _EX specific fields in these structures is not correct. Rigth now we assume that only if WPA is supported that we can access _EX fields. This may not be correct, but safe.
Introduced debug level 5 for spin locks which are most frequently called.

The work added to ndis_irq_bh workqueue should be flushed before the driver
is halted.  Calling flush_schedule_work in NdisFreeBufferPool seemed to work
for Centrino driver. However, that is not correct. It is possible that
between NdisFreeBufferPool and NdisDeregisterInterruput new interrupts can
occur and new work gets scheduled. Right thing to do is to flush workqueue
from NdisDeregisterInterrupt. However, drivers don't exepct interrupt
handler to be called anymore. So we first set ndis_irq->enabled to 0, and
then flush the workqueue. That way ndis_irq_bh function will not call the
interrupt handler and workqueue gets flushed.

TI driver has been crashing the kernel when the card is removed. This is
because the driver never calls NdisDeregisterInterrupt! So we now check
in call_halt if the interrupt has been deregistered or not and if not,
call NdisDeregisterInterrupt explicitly. The TI card doesn't crash anymore.

NDIS_OID_ASSOC_INFO was being called with wrong size. To avoid errors,
various sizes are stored in const ints which are used latre on.

While poisoning fill fields of handle, store different values for each
pointer, so if the driver calls any function in these fill fields, we know
which function (based on the offset) is to be implemented. With this, I
know RT2400 driver calls function at offset 27 in fill3. However, I couldn't
find any documentation on that function, so it is not implemented yet.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@542 388e3070-e2af-4704-926e-a220f121eef5