In wpa_set_key, adding and removing keys are exactly same except for the OID, so no need for remove_key function. This also fixes long time bug with TI driver that crashes when key is removed. Apparently all information about the key being removed should be given.
Return value from WPA support functions should be -1 in case of error, not
-EINVAL.

In wpa_associate, when setting encryption mode, the keys are not configured
yet, so the driver will not set the ecryption mode to what is actually being
set, so don't check for it.

wpa_deauthenticate is same as wpa_disassociate; no need for duplicate code.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@850 388e3070-e2af-4704-926e-a220f121eef5