Loading of drivers has been refactored. Earlier drivers are loaded by userspace loadndisdriver in bits and pieces into ndiswrapper without full error checking. Now a driver is fully loaded by loadndisdriver. It should do all error processing (not sure if _all_ error checking is done right now, but it is much better) and only if everything is valid, should it send the driver with one ioctl to ndiswrapper module.
The new driver uses arrays for system files, devices, bin (firmware) files
and device settings. One side effect of this is that NdisWriteConfiguration
doesn't work any more as it can't add new entries into settings. This has
to be fixed soon (although no drivers seem to write configurations).

There is also lot of indentation changes, so code is in Linux kernel style;
not that this will ever be merged into kernel, but it is easier for others to
submit patches etc.

With a bit more cleanup, it looks like ndiswrapper should be ready for 1.0
release and needs less TLC (I hope).


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@866 388e3070-e2af-4704-926e-a220f121eef5