Access linux spinlock member of union KSPIN_LOCK explicitly. This should fix warnings/errors with old compilers (although both gcc-2.95 and gcc-3.3.5 work fin without it).
Prevent spinlock use if Linux spinlock size is not same as that of Windows
spinlock. This should thwart anyone that removes the check in the Makefile
(on CONIFG_DEBUG_SPINLOCK). The space used by Linux spinlock must be less
than or equal to Windows spinlock, or else kernel will crash.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@902 388e3070-e2af-4704-926e-a220f121eef5