Once scatter/gather DMA list is created, Mdl buffers have to be recreated with info from sg list. This is done in Windows by BuildMdlFromScatterGatherList. Here we do it on the fly in send_packets_sg_dma. The sg list info in the case of sg DMA is now stored in handle instead of packet so we don't have to keep allocating/freeing this info. If sg DMA is requested by the driver but is disabled by ndiswrapper due to performance issues, we create dummy sg list and store it in each packet.
With all these changes, RTL8180L card gives rather bad performance with sg
enabled - I get about 30KB/s with sg enabled and about 640KB/s with sg
disabled sending packets. I am not sure if this is problem with realtek
driver (I assume not, as it must be used in Windows) or problem with
setting up sg list (since otherwise things work, it is difficult to find
out if there is a problem). The other card I have that uses sg DMA is
Centrino 2100, but that miniPCI card doesn't have antenna, so it barely
establishes a connection, not a candidate for performance measuring.

For now sg DMA is disabled (in ndis.h) and put in freezer - unless someone
else sheds some new light, I won't bother with this.

Unnecessary memset'ing buffers/packets to 0 are commented out. If things
work with this, these statements can be removed later.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1038 388e3070-e2af-4704-926e-a220f121eef5