Calling pnp_event_notify in ndis_remove_one causes Marvell chipset to cause kernel crash, so disable it (I don't think any driver benefits with calling pnp_event_notify anyway).
Do extra sanity checking in NdisMDeregisterInterrupt.

With these changes Marvell driver works with removing driver with the
device still inserted. However, for some reason, the card seem to be
left in a bad state, as next time when module is inserted, we get the
infamous "failed with error -22" problem. Ejecting and inserting the
card back makes it work again. Need to check what is causing this
problem.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1115 388e3070-e2af-4704-926e-a220f121eef5