Initialize callback_objects list (otherwise, ntoskernel_exit will cause crash).
In KeWaitForMultipleObjects, put the thread in TASK_INTERRUPTIBLE or
TASK_UNINTERRUPTIBLE state as appropriate before inspecting the objects to
find out which ones need to be waited on. Otherwise, there could be a race
(see the code for details) and the thread may not be woken up in some cases.
It is possible there are still some issues with this: Zydas driver doesn't
work yet and the error message indicates it is related to events!


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1189 388e3070-e2af-4704-926e-a220f121eef5