There are two different kinds of timers: kernel timers and NDIS timers. When a kernel timer is fired, it attains signal event (that may wake up threads waiting on that event) whereas NDIS timers don't.
Then there are  two different kinds of DPCs that can be associated with
those timers: kernel DPCs, which are executed through DPC worker where at
most one DPC is put in the queue when timer is fired and NDIS DPCs - well,
they are not called DPCs by DDK - which execute when timer is fired.
However, within NDIS, there are two different timers, one set with
NdisSetTimer etc. and one set with NdisMSetTimer (so called miniport timers).
Of these, the first kind uses NDIS DPCs and the latter uses kernel DPCs
(one unique DPC can be in the queue).

So all timer/DPC functions set appropriate field and schedule DPC or
execute it when timer is fired.

Use separate lock (kdpc_list_lock) instead of ntoskernel_lock so that
if the DPC calls any kernel function that also needs ntoskernel_lock,
it won't deadlock.

The logic for checking and resetting signaled state is fixed: In the case
of mutex (which is also SynchronizationEvent), the value of signal_state
can be negative if the same thread recursively obtains the mutex. With this
fix, rt2500usb driver no longer hogs the CPU (the thread kept going to sleep
waiting on an event which is already in signaled state so it would wake up
again immediately and would go back to sleep and so on).


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1274 388e3070-e2af-4704-926e-a220f121eef5