Added functions to set rts and fragmentation threshold values.
Call MiniportPnpEventNotify in loader.c itself for USB devices instead of
passing it to ndiswrapper_remove_device.

Added kdpc worker thread for NDIS timer functions. The DPCs for NDIS are
different from that for kernel DPCs: For kenrel DPCs, only one instance
of a DPC can be in the queue whereas for NDIS, no such restriction applies.
Not sure if a separate NDIS kdpc worker is required - executing in the
bh of timer itself works same, but the description for Ndis timers seems
to imply that DPCs are queued and executed later. The cancel timer function
is also changed to reflect this.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1288 388e3070-e2af-4704-926e-a220f121eef5