Added functions to create and remove kthread structures.
Create kthread structure before calling MiniportInitialize and destroy
it after it is done. Do the same for MiniportHalt.

Use usb_submit_irp to submit irp for USB layer in
pdoDispatchInternalDeviceControl. usb_submit_irp makes sure that it
doesn't touch an IRP for which USBD_STATUS_PENDING is returned (such
irp's will be processed in completion function for urb's). Theoritically,
there is still a race condition: If USBD_STATUS_PENDING is returned for
an urb, usb_submit_irp could return this value back to the driver (which
called IoCallDriver). In the meantime (after submitting urb for which
USBD_STATUS_PENDING was returned but before this value is passed back
to the driver), urb may be transferred and completion function could
run, which returns the irp, which may be freed. I can't think of a way
to prevent this race - we can't maintain status of irp in a field to help
us determine if we should return STATUS_PENDING, as the irp may already
have been freed in the completion function.

Set urb/irp status fields as appropriate in various usb layer functions.

With these RT2500 USB driver is happy.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1295 388e3070-e2af-4704-926e-a220f121eef5