Use own workqueue instead of shared workqueue - this doesn't solve anything, but won't interfere with other workers, and may prevent from keyboard lockup etc. This should only be a tempoaray "fix" and everything should work fine even with shared workqueue.
In KeWaitForMultipleObjects, if value of timeout is 0, first check if wait
can be satisfied for all objects without grabbing them. Only if they can
all be grabbed, should we do so in the next step. If going back to wait
for more objects in the loop, set wait_jiffies to 0 if timeout was NULL,
instead of to res, as we set res to 1 after schedule returns.

Threads are notification events, not synchronization events.

With smaller HZ, each timer interrupt occurs at longer intervals, so wait
for one timer interrupt longer (add 1 to SYSTEM_TIME_TO_HZ), else the driver
may not expect to be woken up quickly.

Use tasklet, instead of workqueue to submit URBs. It turns out that ZyDas
driver monopolozies shared workqueue (due to events/threads/etc issue) and
URBs are not submitted any longer. With tasklet, the worker submits URBs
and the driver goes ahead (still wrong for it to monopolize workqueue, but
helps understand what the problem is). Use URB completion function itself
to process completed URBs instead of another workqueue. This again needs to
be revisited when ZyDas driver is fixed.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1310 388e3070-e2af-4704-926e-a220f121eef5