Finally ZyDas driver works (although partially and there are some issues)!
Scheduling NDIS kdpc work through a worker makes ZyDas driver go crazy - it
keeps hogging up the cpu and it looks like that is due to too much of work
items that get added to ndis_kdpc_worker. Fix (temporary?) is to get rid
of worker for this and execute kdpc in the timer handler itself. This has
to be investigated to see what causes this behavior: It could be an error
in canceling timers.

Apparently there is some memory corruption with ZyDas driver: With DEBUG_SLAB
enabled, kernel complains about slab corruption with link_status_handler
and skb_release_data, although it doesn't crash kernel. Fixing this may
resolve other issues with this driver. With ZyDas, if statistics collection
is enabled, some packets take a few seconds for round trip; if they are
disabled, it works fine. Only WEP works - with wpa_supplicant, slab corruption
errors appear and wpa_supplicant itself goes into funk (it goes into an
infinite loop printing something - I can't remember now and it doesn't matter
anyway, at least until slab corruption issues are fixed).

Since ZyDas has open source driver, it is not necessary to get this driver
working with ndiswrapper, but this driver is one heck of a test case: it
(ab)uses events and threads. Getting this driver to work should help fix
bugs in ndiswrapper.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1324 388e3070-e2af-4704-926e-a220f121eef5