Merged all changes that went in HEAD branch.
Timers and DPCs are simplified - until now the major pain of dealing with
kernel timers and NDIS timers, kernel DPCs and NDIS DPCs was due to confusing
explanation for canceling of timers (e.g., KeCancelTimer) which seems to
indicate that DPCs must be dequeued when canceling timer. However, elsewhere
on MSDN, it seems to say KeCancelTimer should return TRUE if timer is pending
and FALSE otherwise. When FALSE is returned, it means either timer is not
pending or DPC is executing (or executed). This has been confirmed with
developers at #reactos. So all that work of keeping timers and DPCs of
kernel and NDIS has been removed. So far the drivers that have been tested
seem to work fine with this change. If any other drivers fail, this fix may
have to be investigated.

It seems Pre-N (Airgo) driver is quite buggy wrt timers: It sets timer with
NdisSetTimer and cancels with NdisMCancelTimer (instead of NdisCancelTimer).
More annoyingly, it initializes a timer with NdisSetTimer before initializing
it with NdisInitializeTimer. So we now check if a timer has been initialized
or not in wrap_set_timer and initialize it if not. There seem to be some
issues with packets/buffers with this driver, too - this has to be investigated.

Semaphore objects are initialized now as NotificationEvent instead of
SynchronizationEvent and when a thread wants to grab it, we check
signal_state and decrement it if it is signaled.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1342 388e3070-e2af-4704-926e-a220f121eef5