Timers and DPCs are simplified - until now the major pain of dealing with kernel timers and NDIS timers, kernel DPCs and NDIS DPCs was due to confusing explanation for canceling of timers (e.g., KeCancelTimer) which seems to indicate that DPCs must be dequeued when canceling timer. However, elsewhere on MSDN, it seems to say KeCancelTimer should return TRUE if timer is pending and FALSE otherwise. When FALSE is returned, it means either timer is not pending or DPC is executing (or executed). This has been confirmed with developers at #reactos. So all that work of keeping timers and DPCs of kernel and NDIS has been removed. So far the drivers that have been tested seem to work fine with this change. If any other drivers fail, this fix may have to be investigated.
It seems Pre-N (Airgo) driver is quite buggy wrt timers: It sets timer with
NdisSetTimer and cancels with NdisMCancelTimer (instead of NdisCancelTimer).
More annoyingly, it initializes a timer with NdisSetTimer before initializing
it with NdisInitializeTimer. So we now check if a timer has been initialized
or not in wrap_set_timer and initialize it if not. There seem to be some
issues with packets/buffers with this driver, too - this has to be investigated.

Semaphore objects are initialized now as NotificationEvent instead of
SynchronizationEvent and when a thread wants to grab it, we check
signal_state and decrement it if it is signaled.

Prism1 USB driver calls NdisSetTimer with due time as 0, which according to
DDK is wrong - minimum delay should be 10 milliseconds. The driver then sets
two timers with 0 delay and expects them to be executed right away.
Scheduling timer which schedules kdpc's with a worker in this case results
in kernel crash. So we check here for this case and execute kdpc right away.
With this fix, prism1 driver works well, even with collecting statistics and
default timeout for query/set info.

2.6.14-rc2 kernel has new wireless extension (version 20), which uses
get_wireless_stats in wireless handler instead of in device structure, so
it is moved there (conditional on version > 19). There is some issue with
wpa_supplicant though: It doesn't get ssid and wpa_ie etc properly. Not sure
if this issue is with wpa_supplicant or ndiswrapper.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1343 388e3070-e2af-4704-926e-a220f121eef5