Allocate NDIS buffers and NDIS packets ignoring maximum number of descriptors requested, except to give warning; allocated descriptors are chained in free list so a new descriptor is fetched from that list, instead of allocating afresh. When a descriptor is freed, put it in free list if number of allocated descriptors is below a limit, otherwise free it back to kernel.
When converting time units to jiffies for expiration timers, add 1, so that
jiffies is rounded up, instead of down.

KePerformanceCounter should set counter to HZ, not to jiffies.

After even more debugging, found out that ZyDas driver crashes when
ASSOCIATION_INFORMATION oid is called. This is probably because we don't
fill the information as expected by the driver. DDK seems to say that at
least some of the fields in there should be initialized with earlier
request or something similar to that. I don't quite understand the
description about it, so for now, we just don't request it for ZyDas driver.
I thought with this wpa_supplicant doesn't work, but thankfully it does.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1354 388e3070-e2af-4704-926e-a220f121eef5