Added support for RNDIS USB devices. Tested with USR5420 device, WEP and WPA-PSK+TKIP.
Two issues fixed: NdisFreeMemory may be called with second argument as
either tag or length. So we can't rely on length field to decide if
memory was allocated with kmalloc or vmalloc. To note which one is used,
we allocate extra space for 'unsigned long' and store the information
there. NdisFreeMemory then inspects this and frees accordingly.

This device has two interfaces, so USB probe function gets called twice;
however, we need to initialize it only once. So we now check if device
has been already initialized before initializing it (again).


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1360 388e3070-e2af-4704-926e-a220f121eef5