Implemented ExAllocatePoolWithTag so it stores the tag indicating if memory is allocated with kmalloc or vmalloc. This tag is used later when freeing memory with ExFreePool. NdisMAllocateMemory functions now use ExAllocatePoolWithTag etc.
Implemented URB functions used by RNDIS.

RNDIS requires more than 10 URBs. Instead of putting a cap on total number
of URBs a device can allocate, a dynamic list is used. When a URB is freed,
it is put back on the list, so next allocation can retrieve it from the list.
However, if a driver allocates more than certain number of URBs, memory is
freed, to avoid hogging too much of memory.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1363 388e3070-e2af-4704-926e-a220f121eef5