A strange "fix" for ZyDas driver crashing when ASSOCIATION_INFO OID is queried. I have no explanation for this fix and it is a bit of shame to even think this is any fix. But it works and I have spent way too much time to fix this, so for now committing it as it is.
Allocate memory for device extension with ExAllocatePoolWithTag instead
of kmalloc, as Atheros USB driver allocates about 500KB for this, which
can only be allocated with vmalloc (ExAllocateMemoryWithTag allcoates
memory with kmalloc or vmalloc depending on size requested). This driver
doesn't work yet: It tries to dereference NULL pointer. Couple of other
functions used by this driver are partially implemented.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1366 388e3070-e2af-4704-926e-a220f121eef5