Corrected two spinlock release calls, one in usb_exit_device and one in io_worker.
Simplified computing ticks (since year 1601) and made ticks_1601 an inline
function for efficiency.

sis162u driver calls NdisMSleep in atomic context when module is removed
without unplugging the device. This is a bug either in sis driver or
ndiswrapper elsewhere. We now check if called in atomic context and use
udelay if it is. This doesn't solve the bug, but doesn't crash kernel
right away and thus helps in debugging.

Removed multicast_list from wrapper_dev structure, as we allocate and
free the list in set_multicast_list function itself.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1370 388e3070-e2af-4704-926e-a220f121eef5