DPC for a timer can set a timer (even the same timer), so using the same spinlock in timer handler and wrap_set_timer will hang. To avoid this, call DPC without holding spinlock, and use separate locks for each timer. Using separate locks for each timer may not be necessary.
ObfDereferenceObject should release ntoskernel_lock spinlock before
calling FREE_OBJECT, which also uses same lock. With these two fixes, SMP
seems to work, at least on a UP machine with SMP and PREEMPT enabled.

With longer interval for HZ, it is possible that many ticks may result in
0 jiffies to wait. Whenever a timeout is used, add 1 to computed jiffies
to avoid this.

Implemented PoCallDriver.

Don't disable statistics if a query fails; this was done as a workaround
for ZyDas driver which never calls NdisMQueryInformationComplete after
returning NDIS_STATUS_PENDING. This may be a bug in ZyDas driver.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1407 388e3070-e2af-4704-926e-a220f121eef5