IoCreateSymbolicLink should return NTSTATUS.
NdisMRegisterDevice should setup symbolic link, so call IoCreateSymbolicLink.

IoCreateDevice is implemented so all the memory needed is allocated in one
chunk. Apparently this is how Windows implements this function.

ALLOCATE_OBJECT is modified to take size instead of type as parameter, as
sometimes (in the case of IoCreateDevice) we need to allocate space that
is different from size of type.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1424 388e3070-e2af-4704-926e-a220f121eef5