Merging changes in HEAD.
Separating all data structures and implementation so ndis layer is different
from underlying wrapper. With this, devices other than NDIS type can be
supported (the aim is to get a bluetooth device that I have, but after this
separation is done, ndiswrapper can be extended easily to support other
types of devices). This process also cleans up data structures and
implementation (IMO) as right now we just dump everything into wrapper_dev,
for example. This process is not complete yet and it will take sometime to
stabilize with this separation.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1504 388e3070-e2af-4704-926e-a220f121eef5