Get rid of wrap_ndis_packet and put the extra fileds used by ndiswrapper in oob_data structure itself.
I notice odd issue with ar5212 driver: Occassionally hangcheck returns true
so card is reset, but the reset doesn't complete and system needs to be
rebooted. This may be due to the way irq handler is being executed. For now
'pending' field (to count how many times isr was run but handler is not yet
run) is not used, as it was until recently.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1567 388e3070-e2af-4704-926e-a220f121eef5