Mm(Map/Unmap)LockedPages shouldn't update mappedsystemva; instead they should just update flags appropriately.
MmBuildMdlForNonPagedPool should update MDL page descriptions. For now,
however, this is not used by any drivers it seems, so this is disabled
and will be enabled after 1.9 is released so it won't be show stopper for
1.9 release.

CACHE_MDL_SIZE should have enough space for CACHE_MDL_PAGES of PFN_NUMBER,
not ULONG (on x86_64 ULONG is 32 bits, which is not enough for storing
PFN_NUMBER).


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1614 388e3070-e2af-4704-926e-a220f121eef5