NdisOpenConfigurationKeyByIndex should return NDIS_STATUS_FAILURE instead of NDIS_STATUS_SUCCESS, as the driver keeps retrieving registry keys as long as it returns NDIS_STATUS_SUCCESS, causing infinite loop. With this, latest RNDIS devices (e.g., USR5421, WRT54GS2 etc) initialize the device. However, I have been getting random kernel crashes during check_capa stage and I haven't been able to track them down. Linux USB core also complains during initialization and removal, with
[17182323.776000] usb 4-3: link qh1-0001/de9e8100 start 0 [1/0 us]
[17182326.944000] usb 4-3: unlink qh1-0001/de9e8100 start 0 [1/0 us]
[17182326.944000] ehci_hcd 0000:00:1d.7: reused qh de9e8100 schedule

Moreover, I notice that URBs 5-9 that are submitted never come back - they
seem to vanish.

When IRP is canceled, don't free URB and call IoCompleteRequest - if device
is removed, driver cancels IRP for which usb_cancel_urb returns failure
(since device disappeared), for which completion handler is called later
at which point we should free URB. With this, this RNDIS device can be
removed safely.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1654 388e3070-e2af-4704-926e-a220f121eef5