Buffalo WLI-U2-KG54-AI device uses RT2500 USB chipset. This driver when sending URB 185 seemed to interefere with tasklet/workqueue strangely. With 2.6.15 (all tests done on UP machine):  * with tasklets:    - preempt enabled, the system hangs    - smp + preempt enabled, system doesn't hang and softirqs are taking up      all CPU (from this, it looks like some softirq is causing hang with      just preempt)    - real-time preempt (RT) patch, everything works fine  * with workqueues initialization succeeds (past 185th URB), _but_    - if work is scheduled with ndis workqueue, scanning (and possibly      others) block urb completion worker to execute    - if work is scheduled with either anonymous workqueue or wrap workqueue,      everything works
What is every stranger is that RT2500 USB driver works fine with another
device (Belkin F5D7050) with either tasklet or workqueue (not all configuration
mentioned above tested).

Higher powers in linx kernel are needed to solve these mysteries, but
I don't want to deal with some of the bozos on lkml about ndiswrapper.

There is yet another twist to this device: It contains Windows driver on
flash so one could just insert it into any Windows box and it "automatically"
installs the driver and client software. This is done through a filter
driver for HID bus driver (the device becomes a storage device if no driver
is present and then installs software the first time). In Linux, the device
when inserted is recognized as input device and after a few seconds becomes
storage device and story ends there. To get this device to start becoming
network device is to run the filter driver when it is in input device mode.
For this, I created a dummy .conf file to load this driver (since this is
filter driver, there is no associated .inf file) and started the driver.
However, it still doesn't cut it; from what I suspect, this driver needs
full PnP IRP handling to work (PnP manager should send
IRP_MN_QUERY_DEVICE_RELATIONS etc) properly.

Thankfully this device has a hardware switch on the side to enable/disable
this insane feature of installing driver (I found this on a review for this
device after much fighting with it for couple of days). With this feature
disabled, it registers as device for RT2500 driver for which .inf is
provided. This works now with the fix mentioned above.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1671 388e3070-e2af-4704-926e-a220f121eef5