Am1772 driver provides its own allocation and free routines for npaged lookaside list, so when deleting the list, free pointers with the free function associated with the lookaside list instead of with ExFree. This fixes crash with Am1772 driver when removing module.
In KeWaitForMultipleObjects, set the event status to 0 before adding
wait blocks to the list (for objects that a thread needs to wait on).
Otherwise an event may arrive before the event status is set properly but
it is already in the list, causing deadlock / crashes.

KeSynchronizeExecution (don't remember which driver uses it) should lock
spinlock disabling irqs (saving irq flags) as it should be synchronized
with the isr (set with IoConnectInterrupt), and the spinlock should've
been initialized by the driver before calling IoConnectInterrupt, so don't
reinitialize it.

With these fixes, Am1772 driver works fine. It occassinoally complains when
KeWaitForMultipleObjects returns. From debug trace, this seems to be due
to a synchronization event being in signalled state when KeWait is called.
Apparently this driver sets a synchronization event when no threads are
waiting, so event stays in signalled state. Right after that a thread
comes to wait on that event. Since it is already in signalled state,
KeWait returns with STATUS_SUCCESS. This driver doesn't like that. From
what I understand, synchronization event should stay signalled until exactly
one thread is woken up (which is current implementation), and not as this
driver wants it.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1706 388e3070-e2af-4704-926e-a220f121eef5