Simplified various objects that use dispatcher header (events, timers, mutexes, semaphores).
Figured out why Am1772 driver barks with NdisWriteErrorLogEntry (mentioned
before but confirmed now): The driver sets synchronization event when no
thread is waiting on it. Right away a thread comes to wait for this event.
This driver expects at that time for this signal to be in non-signalled
state. So the inference is that synchronization events wake up at most
one thread when KeSetEvent is called; before returning, KeSetEvent should
reset the signal so future threads that need this event has to wait for
it to be signalled again. DDK is not exactly clear about this, but since
this is the only driver that has complicated behavior with synchronization
events, I am going by this for now (IIRC ZyDas driver also uses synchronization
events - have to check if that agrees with this behavior).


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1710 388e3070-e2af-4704-926e-a220f121eef5