If a compiler with fastcall support is used, use that, instead of mungling arguments, so as not to waste registers when fastcall function is called. I believe gcc and Windows use same calling convention with fastcall (so the documentation says and ran couple of tests and checked generated assembly code), but haven't tested if ndiswrapper works now.
Apparently kernel also uses fastcall, but differently from gcc's and
Windows: The callee is supposed to clean the stack, but with kernel's version
caller does. This difference may bite me; for now, include all kernel's header
files before including winnt_types.h so that whatever kernel's definition
of fastcall is works for it and whatever we use in ndiswrapper works for us.
I wonder if I should get rid of support for compilers older than gcc 3.4
(where fastcall is not available), so fastcall mess is not needed.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1791 388e3070-e2af-4704-926e-a220f121eef5