64-bit RNDIS driver seems to crash right after NdisSetTimer. While trying to fix this, I thought it my be due to some assumptions about timers in ndiswrapper: The DPC associated with timers is executed in timer_proc itself, instead of scheduling DPC as Windows does. This patch implements queueing timer DPCs. Moreover, NDIS timers are executed with the wrapper functions that in turn execute DPC function. We use DPCs associated with NDIS and miniport timers for wrapping purposes. Moreover, DPCs are run in a tasklet instead of a worker - the description for DPCs indicate they are run as tasklet instead of worker.
We no longer use 'nr_cpu' of kdpc to keep track of if DPC is queued; instead,
'list' is used appropriately for this.

However, as of yet, the cause of crashes with 64-bit RNDIS is not clear
and this rather large patch doesn't help at all. This patch may need to
be reverted / modified later.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1891 388e3070-e2af-4704-926e-a220f121eef5