Remove references to 'version' file; instead, define driver and utils versions in ndiswrapper.h file.
It is now possible to keep utils directory out of tree, in which case,
directory where driver files exist with 'DRIVER_DIR=' option to make
in utils directory (utils and driver share loader.h and ndiswrapper.h files).

When calling Windows function from Linux function, until now we call
lin2win functions, which then adjust the registers and setup stack as
Windows function expects. In this patch, this is done with macros, so the
overhead of anther function is eliminated. However, this apparently doesn't
work, and I don't see why. Until someone can give debug trace that may
show why this doesn't work, this optimization is disabled (with '#if 0')
in ntoskernel.h



git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@1948 388e3070-e2af-4704-926e-a220f121eef5