When unloading driver, free the settings associated with that driver.
Encode size and type of allocation in one 'unsigned long' when allocating
memory.

When ALLOC_DEBUG is defined, keep track of page allocations (with
__get_free_pages). Since we need to add sizeof(alloc_info) to each allocation
in that case, use wrap_get_free_pages which adds the size appropriately;
otherwise, if driver allocates exactly on page boundary, there won't be
space for alloc_info in that allocation.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@2244 388e3070-e2af-4704-926e-a220f121eef5