ZyDAS driver has two issues: Its NdisScheduleWorkItem queues an item that in turn schedules other work into ntos worker thread, so ndis workqueue has to be separate. For now, I am using wrap_wq thread for this, and let all wrapndis work use shared (default) workqueue, but we should use separate worker thread for that too (e.g., collecting stats takes time and in that time, keyboard, which uses shared workqueue, is hung).
Another issue is that, KeSetEvent wakes up a task that was waiting on that
event, but before that task starts execution (and grabs that object), another
task needs to grab an object and does so, depriving the previos task to go
back to wait. Although this doesn't seem to be a problem, this may be a
problem for some other driver (although, no other driver that I know of
has an issue with this). So, we go back to earlier approach of grabbing the
object by object_signalled function and let it wake up the tasks waiting
on it.

free_mdl should free dispatcher_lock, not ntoskernel_lock.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@2317 388e3070-e2af-4704-926e-a220f121eef5