Patch from Pavel Roskin:
It's not easy to compile ndiswrapper for a kernel that is not currently
used or installed (i.e. there is no link in /lib/modules).  Both KVERS
and KBUILD should be set, which is inconvenient.

The logic that has been used in many other standalone drivers is
different.  The version of the current kernel is only used to locate the
kernel tree.  It's possible that the kernel sources have been updated,
so the actual kernel version should be taken from the configured kernel
tree.

Some older kernels were only installing the "source" link, so if the
"build" link is missing, "source" is the fallback.  It's important if
ndiswrapper still aims to support 2.4 kernels.

If .config is missing, it should be an error.  Otherwise, there is a
risk that the module will be miscompiled.

"readlink -f" is used so that the actual tree location is shown, as
opposed to the link.

Now it's sufficient to override KBUILD.  It's still possible to override
KVERS if the kernel is installed.

Signed-off-by: Pavel Roskin <proski@gnu.org>


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@2427 388e3070-e2af-4704-926e-a220f121eef5