To support RT-preempt, sometime back we danced around with spinlocks and cpuvars to make sure an ndiswrapper task is not preempted by another ndiswrapper task (but other tasks can preempt ndiswrapper tasks). That seemed to work for about couple of rt releases but not anymore. At least for now, don't treat preemption at all when rt is used (this will likely break some drivers). Also, when preempt is not used, use warp(ed) preempt already being used.
It seems that not using preempt is problematic with some drivers (acx111 at
least), so for diagnostic purposes, report what kind of preempt is used when
module is loaded.

Since this module is meant for x86 and x86_64 only, let compile fail on other
architectures (some people apparently tried it on ppc / arm).


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@2466 388e3070-e2af-4704-926e-a220f121eef5